virtio

Module gpu

Structs

BackingMemoryEntry

CreateResource2D

CtrlHeader

DisplayInfo

DisplayMode

FlushResource

ResourceAttachBacking

Command structure to attach backing pages to a GPU resource. This header is followed by a number of BackingMemoryEntry structures.

SetScanout

SimpleResourceAttachBacking

TransferToHost2D

Enums

CtrlType

VirtioGpuFormat