Trait virtio::Write

pub trait Write {
    fn write(&mut self, data: &[u8]) -> Result<()>;
    fn flush(&mut self) -> Result<()>;
}

Write bytes to a source.

This trait is similar to the std::io::Write trait, except that this trait is pared down to a minimum and works in a no_std environment.

Required Methods

fn write(&mut self, data: &[u8]) -> Result<()>

Write all bytes in data.

fn flush(&mut self) -> Result<()>

Flush any output buffers, if they exist.