Struct virtio::vsock::packet::Packet

pub struct Packet { /* private fields */ }

Implementations

impl Packet

pub fn new(buffer: Vec<u8>) -> Result<Self>

Creates a new Packet from an existing buffer.

pub fn new_header_only() -> Self

Creates a new control Packet with only a header.

pub fn new_with_payload(payload_len: usize) -> Self

Creates a new data Packet with the given payload length.

pub fn get_src_cid(&self) -> u64

Gets the source CID.

pub fn set_src_cid(&mut self, src_cid: u64)

Sets the source CID.

pub fn get_dst_cid(&self) -> u64

Gets the destination CID.

pub fn set_dst_cid(&mut self, dst_cid: u64)

Sets the destination CID.

pub fn get_src_port(&self) -> u32

Gets the source port.

pub fn set_src_port(&mut self, src_port: u32)

Sets the source port.

pub fn get_dst_port(&self) -> u32

Gets the destination port.

pub fn set_dst_port(&mut self, dst_port: u32)

Sets the destination port.

pub fn get_len(&self) -> u32

Gets the payload length.

pub fn get_type(&self) -> Result<VSockType>

Gets the type of socket the packet is intended for.

pub fn get_op(&self) -> Result<VSockOp>

Gets the op that the packet represents.

pub fn set_op(&mut self, op: VSockOp) -> Result<()>

Sets the op that the packet represents.

pub fn get_flags(&self) -> VSockFlags

Gets the flags.

Only currently used for connection shutdown requests.

pub fn set_flags(&mut self, flags: VSockFlags)

Sets the flags.

Only currently used for connection shutdown requests.

pub fn get_buf_alloc(&self) -> u32

Gets the size of the peer’s stream buffer.

Used to facilitate flow-control calculations.

pub fn set_buf_alloc(&mut self, buf_alloc: u32)

Sets the size of the stream buffer.

pub fn get_fwd_cnt(&self) -> u32

Gets the number of bytes that the peer has read out of the stream buffer (hence making space in the buffer).

Used to facilitate flow-control calculations.

pub fn set_fwd_cnt(&mut self, fwd_cnt: u32)

Sets the number of bytes read from the stream buffer.

pub fn get_payload(&self) -> &[u8]

Gets the payload.

pub fn set_payload(&mut self, data: &[u8]) -> Result<()>

Sets the payload of a data packet from a slice.

The length of the slice must match the packets configures payload length. Only usable if the packet’s op is VSockOp::Rw.